Common: Rename common/events/ -> common/agent_events/
