+++ monkey/common/agent_event_serializers/i_agent_event_serializer.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Dict, List, Union
 
-from common.events import AbstractAgentEvent
+from common.agent_events import AbstractAgentEvent
 
 JSONSerializable = Union[  # type: ignore[misc]
     Dict[str, "JSONSerializable"],  # type: ignore[misc]
