+++ monkey/common/agent_events/credentials_stolen_events.py
